refactor: serverRequest, useRequest분리 , 순환참조 에러개선 및 환경 혼합 문제개선
